requirements.txt
crewai
requests
streamlit
python-dotenv
groq
